+++ src/main/java/blackjack/domain/participants/BettingProfit.java
@@ -1,6 +1,5 @@
 package blackjack.domain.participants;
 
-import blackjack.domain.Profit;
 import java.util.LinkedHashMap;
 import java.util.Map;
