+++ src/test/java/blackjack/domain/BettingTest.java
@@ -7,6 +7,7 @@ import blackjack.domain.participants.Betting;
 import blackjack.domain.participants.Hands;
 import blackjack.domain.participants.Name;
 import blackjack.domain.participants.Player;
+import blackjack.domain.participants.Profit;
 import blackjack.domain.participants.Result;
 import java.util.List;
 import org.assertj.core.api.Assertions;
