+++ src/test/java/blackjack/domain/ProfitTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.participants.Profit;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
